--- src/tabular_trees/lightgbm/__init__.py
@@ -1,3 +1,0 @@
-"""Module containing functionality for dealing with lightgbm models."""
-
-from . import trees
