+++ tests/lightgbm/test_LightGBMTabularTrees.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tabular_trees.lightgbm.trees import LightGBMTabularTrees
+from tabular_trees.lightgbm import LightGBMTabularTrees
 from tabular_trees.trees import BaseModelTabularTrees, TabularTrees
